Set expiration on SSH keys used in training (#26)
Sets an expiration on SSH keys created using gcloud compute ssh. This causes the ssh keys to be cleaned up in the login profile after approximately 1 day. Given that it takes several weeks for the profile to reach the 32kb limit, this should prevent us from running into the limit.